+++ Changelog.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add getTokenBySymbol to Ethereum.
 
+- uniswap.config.ts with Uniswap addresses.
+
+- balancer.config.ts with Balancer addresses
+
 ## [0.1.2] - 2021-07-27
 
 ### Added